Limit querying latest user transactions in Maniphest to recent IDs
Assuming that we do not set $hours to an astronomically high value,
do not rely only on "dateModified > %d" but reduce to a subset of the
latest 0.5 million rows via the row id which is way more performant.

id is the primary key which autoincrements. We can safely assume that
id(X) < id(X+1) <=> dateCreated(X) < dateCreated(X+1)
so we do not miss any recent transactions.

Bug: T366811
Change-Id: I8cf351fe2d160370a6d9e6a45765ff16d7258a1f